Fix a spike in monotonic counter values after a failed openmetrics scrape (#20063)
* AGTMETRICS-177 skip flush first value on error

flush_first_value indicates that new monotonic counters values should
assume previous value of zero and be included in the very next flush,
instead of waiting for another value to compute a diff.

This breaks when scrape fails. We can no longer assume that the
counter just appeared, and reporting it may cause an unwanted spike.
Aggregator retention for previous values is also short lived, so if
few scrapes are missed, even old counters will appear as new to the
aggregator, causing even bigger spikes.

* Changelog

* Preserve use process_start_time on first successful scrape

If the target process starts after the agent, we may have a few failed
scrapes, but the values on first scrape can be fresh enough to use as
is. 65833120d1b1dc380dd2f9b696d68995907aad93